bot/requirements.txt
MetaTrader5==5.0.45
google-generativeai==0.8.3
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
loguru==0.7.2
pydantic==2.9.2
httpx==0.27.2
tenacity==9.0.0
PyYAML==6.0.2
